refactor(Semantic): restrict the semantics of read0
This commits restrict read0 such that it can no longer be used
to read the initial register value after a write0 occurred.
Instead the read0 is going to abort the action in that case.

This restriction does not effect the expressiveness of koika,
as the value could just be read at the beginning of the cycle
and saved into a locally boundy variable until accessed.

This restriction is a necessary buildingblock of the log_irrelevance
proof used in the upcomming hoare logic.